user/requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
SQLAlchemy==2.0.25
geographiclib==2.0
geopy==2.4.1
python-dotenv==1.0.1
Pyrebase4==4.7.1
DateTime==5.5
pycryptodome==3.20.0
